Check output of `clean_metadata()`

Cleaning metadata can take a series of tries. This function helps summarize and explore the metadata for possible patterns which may help find problems.

Usage

check_meta(meta, date = FALSE)

Arguments

meta

Data frame. Recording metadata. Output of `clean_metadata()`.

date

Logical. Whether to summarize output by date (as well as `site_id` and `aru_id`. Default `FALSE`.

Value

A data frame summarizing the metadata by site_id, aru_type, aru_id, and (optionally) by date. Presents the number of files, directories, and days worth of recordings, as well as the minimum and maximum recording times.

Examples

m <- clean_metadata(project_files = example_files)
#> Extracting ARU info...
#> Extracting Dates and Times...

check_meta(m)
#> # A tibble: 10 × 11
#>    site_id aru_type  aru_id     type  n_files n_dirs n_days min_date           
#>    <chr>   <chr>     <chr>      <chr>   <int>  <int>  <int> <dttm>             
#>  1 P01_1   BarLT     BARLT10962 wav         6      3      2 2020-05-02 05:00:00
#>  2 P02_1   SongMeter S4A01234   wav         6      3      2 2020-05-04 05:25:00
#>  3 P03_1   BarLT     BARLT10962 wav         3      3      1 2020-05-06 10:00:00
#>  4 P04_1   BarLT     BARLT11111 wav         6      3      2 2020-05-06 05:00:00
#>  5 P05_1   BarLT     BARLT10962 wav         3      3      1 2020-05-07 05:00:00
#>  6 P06_1   BarLT     BARLT10962 wav         3      3      1 2020-05-09 05:20:00
#>  7 P07_1   SongMeter S4A01234   wav         6      3      2 2020-05-09 05:25:00
#>  8 P08_1   BarLT     BARLT10962 wav         3      3      1 2020-05-11 10:00:00
#>  9 P09_1   SongMeter S4A02222   wav         3      3      1 2020-05-11 05:00:00
#> 10 P10_1   SongMeter S4A03333   wav         3      3      1 2020-05-11 03:25:00
#> # ℹ 3 more variables: max_date <dttm>, min_time <time>, max_time <time>
check_meta(m, date = TRUE)
#> # A tibble: 14 × 10
#>    site_id aru_type  aru_id     type  date       n_files n_dirs n_days min_time
#>    <chr>   <chr>     <chr>      <chr> <date>       <int>  <int>  <int> <time>  
#>  1 P01_1   BarLT     BARLT10962 wav   2020-05-02       3      3      1 05:00   
#>  2 P01_1   BarLT     BARLT10962 wav   2020-05-03       3      3      1 05:20   
#>  3 P02_1   SongMeter S4A01234   wav   2020-05-04       3      3      1 05:25   
#>  4 P02_1   SongMeter S4A01234   wav   2020-05-05       3      3      1 07:30   
#>  5 P03_1   BarLT     BARLT10962 wav   2020-05-06       3      3      1 10:00   
#>  6 P04_1   BarLT     BARLT11111 wav   2020-05-06       3      3      1 05:00   
#>  7 P04_1   BarLT     BARLT11111 wav   2020-05-07       3      3      1 03:25   
#>  8 P05_1   BarLT     BARLT10962 wav   2020-05-07       3      3      1 05:00   
#>  9 P06_1   BarLT     BARLT10962 wav   2020-05-09       3      3      1 05:20   
#> 10 P07_1   SongMeter S4A01234   wav   2020-05-09       3      3      1 05:25   
#> 11 P07_1   SongMeter S4A01234   wav   2020-05-10       3      3      1 07:30   
#> 12 P08_1   BarLT     BARLT10962 wav   2020-05-11       3      3      1 10:00   
#> 13 P09_1   SongMeter S4A02222   wav   2020-05-11       3      3      1 05:00   
#> 14 P10_1   SongMeter S4A03333   wav   2020-05-11       3      3      1 03:25   
#> # ℹ 1 more variable: max_time <time>