Extract and clean ARU metadata from file names

Using regular expressions, metadata is extracted from file names and directory structure, checked and cleaned.

Usage

clean_metadata(
  project_dir = NULL,
  project_files = NULL,
  file_type = "wav",
  subset = NULL,
  subset_type = "keep",
  pattern_site_id = create_pattern_site_id(),
  pattern_aru_id = create_pattern_aru_id(),
  pattern_date = create_pattern_date(),
  pattern_time = create_pattern_time(),
  pattern_dt_sep = create_pattern_dt_sep(),
  order_date = "ymd",
  quiet = FALSE
)

Arguments

project_dir

Character. Directory where project files are stored. File paths will be used to extract information and must actually exist.

project_files

Character. Vector of project file paths. These paths can be absolute or relative to the working directory, and don't actually need to point to existing files unless you plan to use clean_gps() or other sampling steps down the line. Must be provided if project_dir is NULL.

file_type

Character. Type of file (extension) to summarize. Default wav.

subset

Character. Text pattern to mark a subset of files/directories to either "keep" or "omit" (see subset_type)

subset_type

Character. Either keep (default) or omit files/directories which match the pattern in subset.

pattern_site_id

Character. Regular expression to extract site ids. See create_pattern_site_id(). Can be a vector of multiple patterns to match.

pattern_aru_id

Character. Regular expression to extract ARU ids. See create_pattern_aru_id(). Can be a vector of multiple patterns to match.

pattern_date

Character. Regular expression to extract dates. See create_pattern_date(). Can be a vector of multiple patterns to match.

pattern_time

Character. Regular expression to extract times. See create_pattern_time(). Can be a vector of multiple patterns to match.

pattern_dt_sep

Character. Regular expression to mark separators between dates and times. See create_pattern_dt_sep().

order_date

Character. Order that the date appears in. "ymd" (default), "mdy", or "dmy". Can be a vector of multiple patterns to match.

quiet

Logical. Whether to suppress progress messages and other non-essential updates.

Value

Data frame with extracted metadata

Details

Note that times are extracted by first combining the date, date/time separator and the time patterns. This means that if there is a problem with this combination, dates might be extracted but date/times will not. This mismatch can be used to determine which part of a pattern needs to be tweaked.

See vignette("customizing", package = "ARUtools") for details on customizing clean_metadata() for your project.

Examples

clean_metadata(project_files = example_files)
#> Extracting ARU info...
#> Extracting Dates and Times...
#> # A tibble: 42 × 8
#>    file_name  type  path  aru_type aru_id site_id date_time           date      
#>    <chr>      <chr> <chr> <chr>    <chr>  <chr>   <dttm>              <date>    
#>  1 P01_1_202… wav   a_BA… BarLT    BARLT… P01_1   2020-05-02 05:00:00 2020-05-02
#>  2 P01_1_202… wav   a_BA… BarLT    BARLT… P01_1   2020-05-03 05:20:00 2020-05-03
#>  3 P02_1_202… wav   a_S4… SongMet… S4A01… P02_1   2020-05-04 05:25:00 2020-05-04
#>  4 P02_1_202… wav   a_S4… SongMet… S4A01… P02_1   2020-05-05 07:30:00 2020-05-05
#>  5 P03_1_202… wav   a_BA… BarLT    BARLT… P03_1   2020-05-06 10:00:00 2020-05-06
#>  6 P04_1_202… wav   a_BA… BarLT    BARLT… P04_1   2020-05-06 05:00:00 2020-05-06
#>  7 P04_1_202… wav   a_BA… BarLT    BARLT… P04_1   2020-05-07 03:25:00 2020-05-07
#>  8 P05_1_202… wav   a_BA… BarLT    BARLT… P05_1   2020-05-07 05:00:00 2020-05-07
#>  9 P06_1_202… wav   a_BA… BarLT    BARLT… P06_1   2020-05-09 05:20:00 2020-05-09
#> 10 P07_1_202… wav   a_S4… SongMet… S4A01… P07_1   2020-05-09 05:25:00 2020-05-09
#> # ℹ 32 more rows
clean_metadata(project_files = example_files, subset = "P02")
#> Extracting ARU info...
#> Extracting Dates and Times...
#> # A tibble: 6 × 8
#>   file_name   type  path  aru_type aru_id site_id date_time           date      
#>   <chr>       <chr> <chr> <chr>    <chr>  <chr>   <dttm>              <date>    
#> 1 P02_1_2020… wav   a_S4… SongMet… S4A01… P02_1   2020-05-04 05:25:00 2020-05-04
#> 2 P02_1_2020… wav   a_S4… SongMet… S4A01… P02_1   2020-05-05 07:30:00 2020-05-05
#> 3 P02_1_2020… wav   j_S4… SongMet… S4A01… P02_1   2020-05-04 05:25:00 2020-05-04
#> 4 P02_1_2020… wav   j_S4… SongMet… S4A01… P02_1   2020-05-05 07:30:00 2020-05-05
#> 5 P02_1_2020… wav   o_S4… SongMet… S4A01… P02_1   2020-05-04 05:25:00 2020-05-04
#> 6 P02_1_2020… wav   o_S4… SongMet… S4A01… P02_1   2020-05-05 07:30:00 2020-05-05