Create spectrogram from wave file

Functions from Sam Hache to create spectrograms

Usage

soxSpectrogram(
  file.path,
  out.path = "Spectrograms",
  out.app = "test",
  size = list(x = 2000, y = 1000),
  duration = list(start = 0, end = 180),
  sox.file.path
)

Arguments

file.path

Input must be file with its directory (file.path)

out.path

Default will save in Spectrogram folder in the current working directory otherwise specify location (out.path)

out.app

Appends to the start of the output file. Default is 'test'

size

Size of picture in pixels

duration

Duration of recording to plot

sox.file.path

Path to sox file.

Value

Does not return anything, but does create a image file at out.path.

Details

Create spectrogram and save it wherever you want Input must be file with its directory (file.path) Default will save in Spectrogram folder in the current working directory otherwise specify location (out.path) Size specifies the x and y size of the output in pixels must be list Duration sets the duration of spectrogram default is 3 min and is listed in seconds